requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
Jinja2==3.1.2
gunicorn==21.2.0
email-validator==2.1.0.post1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
